eventree and bdaygift
eventree - Given a tree and asked to remove as many edges from the tree
as possible to obtain a forest with the condition that : Each connected
component of the forest should contain an even number of vertices.
Return the number of edges removed.

bdaygift - Calculate expectation of the sum of a series of numbers.
Numbers are added to the sum if a coin flip results in heads.