Closest numbers
Find the pair of numbers in the array with the smallest difference.
There can be multiple pairs with the lowest difference.

Sort the array and keep track of the lowest difference and the two
numbers that made up that pair. If you find a number with the same
difference, add the pair to the list. If you find a pair with a lower
difference, clear the list, add the new pair, and make this value now
the lowest difference.