Jim and the Orders
Given a list of orders. Each order has a time t that it starts and takes
d time to process. Print the orders in the order that they finish. If
two orders finish at the same time, the one listed in the input first
takes priority.

Create and Order class to keep track of the order number as well as the
start and duration time. Create an array of orders and sort them by what
time they finish. The CompareTo function handle the case where two
orders finish at the same time. Iterate through the sorted array
printing out the number of each order.