Priyanka and Toys
Given a list of N toys with weight n'. If you buy the toy n', you get
all the toys in the range of weight n' to n' + 4. Determine the number
of toys you need to buy to get all the toys.

Sort the list of toys by weight. Iterate over the toys starting with the
least weight. Keep track of the weight of the last toy you bought (I
used Int.MinValue for the case where you haven't bought a toy yet). If
this toy is within the weight range of the last toy you bought, don't
buy this toy. If it's not, you need to buy the toy and update the weight
of the last toy you've bought with this item's weight.