Game of Thrones - I
Given a string S, determine if there if an anagram of S that is a
palindrome.

Iterate through the string and keep track of the number of occurrences
of each letter.

Corner Case: words consisting of only single letter (can have multiple
occurrences) are always palindromes

If the number of letters in the word is EVEN:
The word has a palindrome if no letters have an odd number of
occurrences.

If the number of letters in the word is ODD:
The word has a palindrome only if one letter has an odd number of
occurrences. (e.x. abcdcba where d is 1 occurrence or abcbcba where b
has 3 occurrences)